Renamed fields in protobuf messages
It would appear that the Protobuf compiler enforces Google's C++ naming
conventions, which are that field getter and setter functions should
have lowercase names.  In other words, if my message has a field named
fooBar, then the corresponding setter would be named set_foobar().  This
commit brings the field names in line with that convention, so we don't
have to deal with two different names for one thing.