Added new "published" state
As with the "ready" state, all slaves must now report being in the
"published" state before being allowed to continue.  The purpose is to
be able to send them a TERMINATE command if one or more other slaves are
for some reason unable to complete their timestep in a timely manner so
that the simulation needs to be aborted or restarted.