Merged value and reference sequences
Previously, when using dsb::sequence::Sequence<T>, one had to choose
upfront whether T should be a value or reference type.  This makes it
ideal for return values, but less so for parameters.  (A function should
most of the time not care whether it iterates a sequence by reference or
value.)  This removes Sequence<T&> types, and makes Sequence::Next()
always return T&.  This requires caching in sequences which are generated
on the fly.