Added new class for ExecutionAgent private data
Problem: ExecutionAgent has some internal data and methods that must be
accessible to the execution state objects, such as the list of slaves
and the function to change states.  Having this functionality publicly
available outside the ExecutionAgent, however, is just poor
encapsulation.

Solution: Define a new type, ExecutionAgentPrivate, which contains the
ExecutionAgent's private data.  This is made private, and thus generally
not exposed to the outside world, except to the execution state objects,
to which we explicitly pass a reference.