Removed zmq::context_t from our public API
DSB users shouldn't have to deal with ZMQ stuff, and in fact, it is not
a necessity.  The contexts can be managed internally by DSB, and we can
in fact have as many contexts as we like, as long as sockets which are
supposed to talk to each  other over INPROC channels use the same
context.