Shut everything down if there is no communication
With this commit we've implemented a communications timeout feature,
where the execution broker and the slave providers commit suicide if
there is no communication going on for a pre-set amount of time (which
defaults to 1 hour).  This is to clean up system resources when parts of
the simulation go the way of the dodo.