Remove dsb::fmilib::Fmu1::AdditionalDllDirectory
The class being removed here was introduced to fix VIPROMA-42.  This
fix, as it turned out, was somewhat problematic.  See VIPROMA-67 for
more info.

However, it may still have a role to play as an *alternative* to the
PATH-modifying mechanism with which it was recently replaced, so I am
removing it in this separate commit so it is easy to bring it back.