Fix VIPROMA-78: Assertion failure on termination
This simply ensures that the callback in SlaveControlMessengerV0 is
called with an "aborted" code if the slave is busy when the termination
command is issued. I've also documented this fact.

It also ensures that the PendinSlaveControlConnection is closed on
termination, which has the same effect for connections which have not
been established yet.