Speed up model configuration file parsing
This fixes VIPROMA-81.  The problem was that the variable info was
stored in a std::vector and looked up via a linear search for each
connection.  For M variables and N connections, the complexity was
therefore O(M*N).

The solution was to cache the variable info in a std::map instead, which
speeds things up 600x or so.