Ensure unittests run on Linux
When the same port number is used two times in quick succession (even
if the socket is explicitly closed in between), the bind operation
sometimes fails with an "Address already in use" error.  I don't know
whether this is a Linux thing, a ZeroMQ thing, or a result of the fact
that I'm running Linux in a virtual machine.