More robust unittests
Some of the unittests were reusing the same endpoint addresses over and
over again, which resulted in sporadic "Address already in use" errors
on Linux, even when the sockets were closed explicitly.

This appears to be related to the fact that ZeroMQ does all its internal
socket management and bookkeeping in background threads, so even if a
socket appears to have been closed successfully, the internal cleanup
work may still be ongoing, and the address may not be ready for reuse
yet.

Furthermore, the dsb_comm.P2PReqRepProxied test would sometimes stall
because a message simply did not get routed correctly through the proxy.
The reason was that we tried to connect a new client to the proxy with
the *same* identity as a previous one, before the proxy had even
discovered that the previous one had disconnected.  Consequently, the
new connection was ignored and the messages did not get routed.
Solution: Use a new identity.

I don't know why these problems have mainly surfaced on Linux.  It could
be that the unittests run faster there than on Windows, or that ZeroMQ's
background work runs slower.