Never destroy the global ZMQ context
Using a global zmq::context_t object in a DLL is a problem due to the
order in which Windows performs various operations -- in particular
destruction of static data -- during DLL unloading/shutdown.  See these
links for more info:

    https://github.com/zeromq/libzmq/issues/1144
    http://stackoverflow.com/q/19795245

While the issue above is reported to be fixed, it appears to just have
shifted the problem elsewhere.  The solution for us seems (for now) to
be that we simply never destroy the object.  The OS will reclaim all
resources on program termination anyway.