Assign a user-specified name to each slave
We've always assigned a name to each slave in dsbexec's configuration
file, but the name was only used in the configuration parsing stage,
as a convenient alias for the numeric slave IDs.  Now, the names become
an integral part of the execution structure, and will in the future be
transmitted to observers etc.  Note that this commit doesn't actually
*use* the names for anything yet, it only sets up the infrastructure so
they are passed to, and stored in, the ExecutionManager.

This fixes VIPROMA-98.