Fix bugs with step timeouts in dsbexec
The value of step_timeout_multiplier is now checked during configuration
parsing to see if it would result in a timeout less than 1 ms.  In
addition, the timeout was for some reason reduced to a resolution of
*seconds* before being passed to Step().

This is kind of a workaround for VIPROMA-101, or at least it prevents
the dsbexec user from running into that issue.