Maintain information about slaves in execution
Until now, detailed information about the type of each slave (such as
type, variables, description, etc.) had to be obtained from slave
providers.  The slaves themselves did not provide this information, and
so there was no easy way to access it in the context of an execution.

In particular, the ExecutionManager can make good use of such
information to verify the validity of connections and so on, and
observers absolutely REQUIRE that info.

This commit adds functionality whereby the master can ask each slave for
such information, and does so. This information is acquired just after a
slave has been added, and is stored in the ExecutionManager.

This fixes most of VIPROMA-97.