Allow default slave names
Previously, a slave name had to be given to AddSlave(). Sometimes,
however, it may be convenient to let DSB simply choose one.  If the name
is empty, DSB will now use "_slaveXX" by default, where XX is the
numeric ID.  Since user-supplied names can't start with an underscore,
and the numeric IDs are guaranteed to be unique, the names will be too.