Improve the P2P sockets
Firstly, I removed the requirement of a strict send-receive-send-receive
order on P2PReqSocket operations, simply because there is no *technical*
reason to have it, and there are a few use cases for *not* having it
(such as if the previous request timed out).

Secondly, and somewhat related, I added an Ignore() method to
P2PRepSocket, which allows the server to ignore the last received
request (e.g. if it is invalid).  Combined with the previously mentioned
change, the P2P sockets may now be used in a fully asynchronous manner.

Thirdly, these changes allowed me to remove the handwritten move
constructors/assignment operators in favour of default ones.

Finally, I added the workaround for ZeroMQ issue 1264 to wherever the
underlying ZMQ sockets get closed in these classes, so it never becomes
a problem when using them.  This also allowed me to remove it from
dsb_bus_slave_control_messenger_v0.cpp.  (See also VIPROMA-25.)