Add Wait*() and 'send more' flag to dsb::comm
This removes the Receive() overload that takes a timeout and replaces it
with a separate WaitForIncoming() function to use with the plain
Receive().  This has the benefit of separating concerns, plus
WaitForIncoming() can also be used together with plain
zmq::socket_t::recv() and even with P2P*Socket::Receive().

I've also added a corresponding WaitForOutgoing() function which can be
used for sending messages with a timeout.

Finally, I added a 'more' flag to Send(), so it can be used to enqueue
parts of multiframe messages.  This is the first time I use a scoped
enum in DSB, so for future convenience I made a macro in config.h for
generating bitwise operators for such types.