Warn about unconnected vars, disable by default
Warnings are now disabled by default (even the ones that were there from
before), and must be explicitly enabled with the -w switch.  If this
switch is used, the old warnings are printed, along with a new one for
each input or output variable which does not participate in a
connection.