Add dsb::async::CommThread
This is an abstraction of the pattern used in dsb::domain::Controller
and dsb::execution::Controller, where a background communications thread
is controlled by an object in the foreground thread.

This does away with the ZMQ-based "inproc RPC" mechanism in favour of
more "standard" C++ facilities, only using ZMQ to notify the background
thread that a task is waiting.