Bye, bye broker (and more - much, much more)
This takes the broker out of the slave-to-slave communication too, thus
finalising its removal.  DSB is now a fully distributed system with
automatic discovery and no central point of communication.  This fixes
VIPROMA-113.

In the process, I have done some more long-needed renovation and
improvement of various parts of DSB, and fixed some other
long-standing issues. These are:

  - VIPROMA-47
  - VIPROMA-52 (obsoleted by the removal of broker)
  - VIPROMA-73
  - VIPROMA-79
  - VIPROMA-80
  - VIPROMA-85
  - VIPROMA-87 (obsoleted by the removal of broker)