Improve slave error handling
This adds a distinction between fatal and non-fatal errors on the slave
side, where the former are errors that do not prevent continued
simulation (such as failure to set a variable value due to it being out
of range).  Currently, all errors are still treated as fatal by
coralmaster, but now it at least has the option of making this
distinction in the future.