Allow master to set variable recv. timeout
The slave's variable value communication timeout was introduced in
commit 0478d68, but only as a setting internal to the slave (and
actually hard-coded in SlaveAgent).  Now, it can be set by the master.

As this made for yet another parameter in the Execution and
ExecutionManager constructors, I've introduced a new struct called
ExecutionOptions to hold this and other existing parameters, as well as
others which I expect to add in the future.