Fix bug in ProviderCluster::GetSlaveTypes()
master::ProviderCluster::GetSlaveTypes() calls HandleGetSlaveTypes(),
which uses bus::SlaveProviderClient::GetSlaveTypes() to get information
about slave types from slave providers.  For the sake of efficiency, the
latter will cache the information, so additional calls don't require
further network communication.  This has an effect on when the
completion handler gets called.  If the information is not cached, it
gets called later, as a result of a reactor event (the slave provider
replies), whereas if it is cached, it gets called immediately.
Unfortunately, this messed with the remainingReplies counter in
HandleGetSlaveTypes(), so it was decremented before it had been
incremented, leading to a negative value and a failure to detect when
all slave type information had been received.