Fix issue where slave types are not cached
`SlaveProviderClient` is supposed to cache the list of slave types it
receives from a slave provider, so it doesn't have to send a new request
the next time `GetSlaveTypes()` is called.

However, the slave type list was assigned to a *local* variable
(`slaveTypes`) rather than the appropriate member variable
(`m_slaveTypes`), while still setting the m_slaveTypesCached variable to
`true`.  This made it appear as though the slave types had been cached,
causing `GetSlaveTypes()` to return an empty list the next time it was
called.